Create build-zip.yml
Introduce CI/CD workflow for on-demand ZIP build using 10up/action-wordpress-plugin-build-zip. This update implements a GitHub Action to automate the creation of a ZIP file for our WordPress plugin when it was run manually on GitHub, GitHub CLI, or the REST API.

The ZIP is available for download immediately, facilitating rapid testing and deployment. To optimize storage, the ZIP file is retained for only 24 hours after creation. This ensures our CI/CD pipeline remains efficient and aligned with our development practices.